+++ packages/collection_editor/vite.config.ts
@@ -16,7 +16,7 @@ export default defineConfig(() => ({
             name: "FireCMS collection editor",
             fileName: (format) => `index.${format}.js`
         },
-        target: "esnext",
+        target: "ESNEXT",
         sourcemap: true,
         rollupOptions: {
             external: isExternal