+++ packages/mongodb/vite.config.ts
@@ -18,7 +18,7 @@ export default defineConfig(() => ({
             name: "FireCMS PRO",
             fileName: (format) => `index.${format}.js`
         },
-        target: "esnext",
+        target: "ESNEXT",
         sourcemap: true,
         rollupOptions: {
             external: isExternal