+++ CHANGELOG.md
@@ -86,7 +86,7 @@
         ],
         build: {
             modulePreload: false,
-            target: "esnext",
+            target: "ESNEXT",
             cssCodeSplit: false,
         }
     })